% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRownoColnoCellpartnoSparqlQuery.R
\name{GetRownoColnoCellpartnoSparqlQuery}
\alias{GetRownoColnoCellpartnoSparqlQuery}
\title{SPARQL query for dimensions in RDF data cube}
\usage{
GetRownoColnoCellpartnoSparqlQuery(forsparqlprefix)
}
\arguments{
\item{forsparqlprefix}{PREFIX statements for SPARQL query}
}
\value{
SPARQL query
}
\description{
SPARQL query for dimensions in RDF data cube
}

