% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTwoDimTableFromQb.R
\name{GetTwoDimTableFromQb}
\alias{GetTwoDimTableFromQb}
\title{Get two dimensional table as data.frame from RDF data cube}
\usage{
GetTwoDimTableFromQb(store, forsparqlprefix, domainName, rowdim, coldim,
  rowlabeldim = NULL, collabeldim = NULL)
}
\arguments{
\item{forsparqlprefix}{PREFIX part of SPARQL query}

\item{domainName}{domainName for the RRDF cube}

\item{rowdim}{vector of dimension and attribute names for row identification}

\item{coldim}{vector of dimension and attribute names for column identification}

\item{rowlabeldim}{vector of dimension names to define the row label data.frame}

\item{collabeldim}{vector of dimension names to define the column label data.frame}
}
\value{
data.frame with result of query, the observations, with
attributes for rowLabelDf, colLabelDF, rowlabelURI, collabelURI,
obsURI
}
\description{
Get two dimensional table as data.frame from RDF data cube
}
\details{
The present code uses R to create the two dim representation.
This could also be done using a SPARQL query. Most likely this would be a more
clear approach.
}

