% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeHTMLfromQb.R
\name{MakeHTMLfromQb}
\alias{MakeHTMLfromQb}
\title{Make HTML table representing RDF data cube}
\usage{
MakeHTMLfromQb(store, forsparqlprefix, dsdName, domainName, dimensions, rowdim,
  coldim, idrow, idcol, htmlfile = NULL, useRDFa = TRUE,
  compactDimColumns = TRUE, showProcedure = TRUE, debug = FALSE)
}
\arguments{
\item{store}{RDF data store containing cube}

\item{rowdim}{Row dimensions}

\item{coldim}{Column dimensions}

\item{idrow}{idrows}

\item{idcol}{idcols}

\item{htmlfile}{path to file with HTML}

\item{useRDFa}{if TRUE include RDFa markup (default)}

\item{compactDimColumns}{if TRUE compact dimension columns and add pretty header (default)}

\item{showProcedure}{If TRUE show in each row the projection procedurevalue}

\item{debug}{If TRUE give debug information}
}
\value{
path to file with HTML
}
\description{
Make HTML table representing RDF data cube
}

