% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PresentQbAsHtml.R
\name{PresentQbAsHtml}
\alias{PresentQbAsHtml}
\title{Create HTML file for RDF data cube given by turtle file}
\usage{
PresentQbAsHtml(dataCubeFile, htmlfile, rowdim, coldim, idrow, idcol)
}
\arguments{
\item{dataCubeFile}{}

\item{htmlfile}{}

\item{rowdim}{}

\item{coldim}{}

\item{idrow}{}

\item{idcol}{}
}
\description{
Create HTML file for RDF data cube given by turtle file
}

